Very Basic Camp Model
Implemented a very basic model for the campsites (it's just a triangular
prism). Double checked the functionality of the building randomise in
preparation for merge. It should be noted that the 'roads' in the level
are simply for visualisation purposes. They have no colliders and should
not interfere with anything.