Car and Misc Pickups
Added scripts to all car and misc pickups. Created a script for the
player that tracks how much the car parts they are carrying are worth.
Linked this script with the Mechanic AI and the Maintenance script so
the player will return the parts to the mechanic when in range.